requirements.txt
Django==5.0
psycopg2-binary==2.9.10
gunicorn==21.0.0
python-dotenv==1.0.0